Change recent gas controller to recent block controller
